% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cvContribution_CNV_MCCA}
\alias{cvContribution_CNV_MCCA}
\title{CNV Contributions from Sparse mCCA}
\format{A vector of length 558}
\source{
\url{http://dlin.web.unc.edu/software/iboost/}
}
\usage{
cvContribution_CNV_MCCA
}
\description{
A vector of contributions for copy number variation 
from the cross validated Sparse mCCA analysis.
}
\keyword{datasets}
