% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{foldMem}
\alias{foldMem}
\title{Fold membership}
\format{An object of class \code{integer} of length 558.}
\source{
\url{http://dlin.web.unc.edu/software/iboost/}
}
\usage{
foldMem
}
\description{
A vector of fold memberships of length 558
}
\keyword{datasets}
