% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fullContribution_miRNA_MCCA}
\alias{fullContribution_miRNA_MCCA}
\title{Full miRNA Contributions from Sparse mCCA}
\format{A vector of length 558}
\source{
\url{http://dlin.web.unc.edu/software/iboost/}
}
\usage{
fullContribution_miRNA_MCCA
}
\description{
A vector of contributions for micro RNA
from the full Sparse mCCA analysis.
}
\keyword{datasets}
