% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fullContribution_RNA_MCCA}
\alias{fullContribution_RNA_MCCA}
\title{Full RNA Contributions from Sparse mCCA}
\format{A vector of length 558}
\source{
\url{http://dlin.web.unc.edu/software/iboost/}
}
\usage{
fullContribution_RNA_MCCA
}
\description{
A vector of contributions for gene expression  
from the full Sparse mCCA analysis.
}
\keyword{datasets}
