% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMOData.R
\name{getMOData}
\alias{getMOData}
\title{Load Multi Omics Data}
\usage{
getMOData(dataTypes, cancerType)
}
\arguments{
\item{dataTypes}{As many as desired from "Clinical","CNV","Gene","miRNA","Module","Mutation","OS","Protein"}

\item{cancerType}{Desired Cancer type from "BLCA","BRCA","COAD","HNSC","KIRC","LUAD","LUSC"}
}
\value{
Returns a list of data matrices in the order specified by \code{dataTypes} with common subjects having the specified \code{cancerType}
}
\description{
Load data for a specified cancer and selected data types
}
