% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeContributionPlotObject.R
\name{makeContributionPlotObject}
\alias{makeContributionPlotObject}
\title{Create Contribution Plot Object}
\usage{
makeContributionPlotObject(fullContributions, cvContributions, foldMem,
  scaleType = "SD")
}
\arguments{
\item{fullContributions}{List of unscaled "score vectors" for the analysis using all subjects.  The length of this list should be equivalent to the number of matrices that were used in the analysis and the length of each element should be n.}

\item{cvContributions}{List of unscaled "score vectors" for the cross validation analysis.  The length of this list should be equivalent to the number of matrices that were used in the analysis and the length of each element should be n.}

\item{foldMem}{Vector of length n indicating in which fold the subjects had the scores calculated}

\item{scaleType}{Value of the type of scaling desired.  If the contributions are not clustered together, then "MAD" should be used instead of the default "SD".  See vignette for more details. \cr
"SD" - Standard Deviation \cr
"MAD" - Median Absolute Deviation#'}
}
\value{
An object which contains standardized scores for both the full and CV analysis as well as the fold membership
}
\description{
To standardize and adjust contributions from a cross validation multi-omics analysis where the contributions need to be standardized and weighted to have similar results with the full analysis.
}
\examples{
makecontributionPlotObject(fullContributions,cvContributions,foldMem)
}
