% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMovieObject.R
\name{makeMovieObject}
\alias{makeMovieObject}
\title{Create MOVIE Object}
\usage{
makeMovieObject(fullContributions, cvContributions, foldMem,
  scaleType = "SD", scaleFold = FALSE)
}
\arguments{
\item{fullContributions}{List of unscaled contributions for the analysis using all subjects.  The length of this list should be equivalent to the number of matrices that were used in the analysis and the length of each element within the list should be the number of samples.}

\item{cvContributions}{List of unscaled contributions for the cross validation analysis.  The dimensions of this list should be the same as the `fullContributions`.}

\item{foldMem}{Vector of length equal to the number of samples, indicating in which fold the subjects had their contributions calculated}

\item{scaleType}{Value of the type of scaling desired.  If the contributions are not clustered together, then "MAD" should be used instead of the default "SD".  See vignette for more details. \cr\cr
"SD" - Standard Deviation \cr
"MAD" - Median Absolute Deviation#'}

\item{scaleFold}{TRUE or FALSE on whether or not scaling should be done within Fold. Default is FALSE}
}
\value{
A `movie` object which contains the standardized contributions for both the full and CV analysis as well as the fold membership which will be used for plotting.
}
\description{
To scale and flip contributions from a cross validation multi-omics analysis in order to be align with the contributions from the full analysis.
}
