% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSamplePlotObject.R
\name{makeSamplePlotObject}
\alias{makeSamplePlotObject}
\title{To standardize and adjust "scores" from a cross validation multi-omics analysis where the "scores" need to be standardized and weighted to have similar results with the full analysis.}
\usage{
makeSamplePlotObject(fullScores, cvScores, idMem)
}
\arguments{
\item{fullScores}{List of unscaled "score vectors" for the analysis using all subjects.  The length of this list should be equivalent to the number of matrices that were used in the analysis and the length of each element should be n.}

\item{cvScores}{List of unscaled "score vectors" for the cross validation analysis.  The length of this list should be equivalent to the number of matrices that were used in the analysis and the length of each element should be n.}

\item{idMem}{Vector of length n indicating in which fold the subjects had the scores calculated}
}
\value{
An object which contains standardized scores for both the full and CV analysis as well as the fold membership
}
\description{
To standardize and adjust "scores" from a cross validation multi-omics analysis where the "scores" need to be standardized and weighted to have similar results with the full analysis.
}
\examples{
makeSamplePlotObject(fullScores,cvScores,idMem)
}
