% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMoviePlotGrid.R
\name{plotMoviePlotGrid}
\alias{plotMoviePlotGrid}
\title{Grid of contribution Plots}
\usage{
plotMoviePlotGrid(movieObject, plotType)
}
\arguments{
\item{movieObject}{Output of makeContributionPlotObject()}

\item{plotType}{Specifies which plots to make \cr\cr
\code{"CV"} - CV contribution/Score Plot - (k-1) X (k-1) upper triangular grid where $plot_{i,j}$ Corresponds to the contribution/score plot for matrix j+1 (x-axis) and i (y-axis)\cr\cr
\code{"Full"} - Full contribution/Score Plot - Same gridded plot as above but now for the full scores i.e. the scores from the full analysis\cr\cr
\code{"Comparison"} - Matrix Specific Score Plots - k plots (k is the number of matrices in the analysis) which plot the scores for the CV analysis vs. the scores for the full analysis\cr\cr}
}
\description{
This function creates a grid of plots to help validate the effectiveness and the extent of over fitting for multi-omics methods.
}
