% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSamplePlotGrid.R
\name{plotSamplePlotGrid}
\alias{plotSamplePlotGrid}
\title{Grid of Sample Plots}
\usage{
plotSamplePlotGrid(sampleObject, plotNum)
}
\arguments{
\item{sampleObject}{Output of makeSamplePlotObject()}

\item{plotNum}{Number which indicates which plots to make \cr\cr
1.) CV Sample/Score Plot - (k-1) X (k-1) upper triangular grid where $plot_{i,j}$ Corresponds to the sample/score plot for matrix j+1 (x-axis) and i (y-axis)\cr\cr
2.) Full Sample/Score Plot - Same gridded plot as above but now for the full scores i.e. the scores from the full analysis\cr\cr
3.) Matrix Specific Score Plots - k plots (k is the number of matrices in the analysis) which plot the scores for the CV analysis vs. the scores for the full analysis\cr\cr}
}
\description{
This function creates a grid of plots to help validate the effectiveness and the extent of over fitting for multi-omics methods.
}
