% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSamplePlots.R
\name{plotSamplePlots}
\alias{plotSamplePlots}
\title{Create Sample Plots}
\usage{
plotSamplePlots(sampleObject, plotNum, xAxisPlot, yAxisPlot, colorVar = NULL)
}
\arguments{
\item{sampleObject}{Object generated from \code{makeSamplePlotObject()}}

\item{plotNum}{See plotSamplePlots()}

\item{xAxisPlot}{Index for which matrix should lie on the x-axis}

\item{yAxisPlot}{Index for which matrix should lie on the y-axis}

\item{colorVar}{Variable by which points can be colored by. If NA then no color will be given}
}
\value{
Specified plot.
}
\description{
To create an individual plot from the grid produced by \code{plotSamplePlots()}
}
\examples{
plotSamplePlots(sampleObject,plotNum,xAxisPlot,colorVar)
}
