% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.movie}
\alias{plot.movie}
\title{Make Contribution Plot}
\usage{
\method{plot}{movie}(movieObject, plotType, xAxisPlot, yAxisPlot,
  colorVar = NULL, grid)
}
\arguments{
\item{movieObject}{Object of class ContributionPlotObject}

\item{plotType}{Specifies the type of plot to be generated
\code{"CV"} - CV Contribution/Score Plot - (k-1) X (k-1) upper triangular grid where $plot_{i,j}$ Corresponds to the Contribution/score plot for matrix j+1 (x-axis) and i (y-axis)\cr\cr
\code{"Full"} - Full Contribution/Score Plot - Same gridded plot as above but now for the full scores i.e. the scores from the full analysis\cr\cr
\code{"Comparison"} - Matrix Specific Score Plots - k plots (k is the number of matrices in the analysis) which plot the scores for the CV analysis vs. the scores for the full analysis\cr\cr
\code{"SideBySide"} - Side by side plot of the CV and Full Plots}

\item{xAxisPlot}{Index for which matrix should lie on the x-axis.  This is only used if \code{Grid=F}}

\item{yAxisPlot}{Index for which matrix should lie on the y-axis.  This is only used if \code{Grid=F}}

\item{colorVar}{Variable by which points can be colored by. If NA then no color will be given. This is only used if \code{Grid=F}}
}
\description{
Function utilizing plot which makes a contribution plot
}
