% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.movie}
\alias{plot.movie}
\title{Make Contribution Plot}
\usage{
\method{plot}{movie}(movieObject, plotType, xAxisPlot, yAxisPlot,
  colorVar = NULL, grid, colorVarLabel = NULL)
}
\arguments{
\item{movieObject}{Object of class `movie`}

\item{plotType}{Specifies the type of plot to be generated
\code{"CV"} - CV Analysis Contribution Plot \cr\cr
\code{"Full"} - Full Analysis Contribution Plot \cr\cr
\code{"Comparison"} - Matrix Specific Score Plots \cr\cr
\code{"SideBySide"} - Side by side plot of the CV and Full Plots - Not available if \code{Grid=T}.}

\item{xAxisPlot}{Index for which matrix should lie on the x-axis.  This is only used if \code{Grid=F}.}

\item{yAxisPlot}{Index for which matrix should lie on the y-axis.  This is only used if \code{Grid=F}.}

\item{colorVar}{Variable by which points can be colored by. If NA then no color will be given. This is only used if \code{Grid=F}}
}
\description{
Function utilizing plot which makes a contribution plot
}
